+++ packages/core/@generated/graphql/index.ts
@@ -600,39 +600,6 @@ export type Filter_FacetsFragment = {
   }>
 }
 
-export type SearchSuggestionsQueryQueryVariables = Exact<{
-  term: Scalars['String']
-  selectedFacets: InputMaybe<Array<IStoreSelectedFacet> | IStoreSelectedFacet>
-}>
-
-export type SearchSuggestionsQueryQuery = {
-  search: {
-    suggestions: {
-      terms: Array<{ value: string }>
-      products: Array<{
-        slug: string
-        sku: string
-        name: string
-        gtin: string
-        id: string
-        brand: { name: string; brandName: string }
-        isVariantOf: { productGroupID: string; name: string }
-        image: Array<{ url: string; alternateName: string }>
-        offers: {
-          lowPrice: number
-          offers: Array<{
-            availability: string
-            price: number
-            listPrice: number
-            quantity: number
-            seller: { identifier: string }
-          }>
-        }
-      }>
-    }
-  }
-}
-
 export type ProductDetailsFragment_ProductFragment = {
   sku: string
   name: string
@@ -869,6 +836,48 @@ export type ProductsQueryQuery = {
       }>
     }
   }
+}
+
+export type SearchSuggestionsQueryQueryVariables = Exact<{
+  term: Scalars['String']
+  selectedFacets: InputMaybe<Array<IStoreSelectedFacet> | IStoreSelectedFacet>
+}>
+
+export type SearchSuggestionsQueryQuery = {
+  search: {
+    suggestions: {
+      terms: Array<{ value: string }>
+      products: Array<{
+        slug: string
+        sku: string
+        name: string
+        gtin: string
+        id: string
+        brand: { name: string; brandName: string }
+        isVariantOf: { productGroupID: string; name: string }
+        image: Array<{ url: string; alternateName: string }>
+        offers: {
+          lowPrice: number
+          offers: Array<{
+            availability: string
+            price: number
+            listPrice: number
+            quantity: number
+            seller: { identifier: string }
+          }>
+        }
+      }>
+    }
+  }
+}
+
+export type TopSearchSuggestionsQueryQueryVariables = Exact<{
+  term: Scalars['String']
+  selectedFacets: InputMaybe<Array<IStoreSelectedFacet> | IStoreSelectedFacet>
+}>
+
+export type TopSearchSuggestionsQueryQuery = {
+  search: { suggestions: { terms: Array<{ value: string }> } }
 }
 
 export type ValidateSessionMutationVariables = Exact<{