+++ packages/core/src/components/common/SearchInput/SearchInput.tsx
@@ -1,17 +1,13 @@
-import {
-  formatSearchState,
-  initSearchState,
-  sendAnalyticsEvent,
-} from '@faststore/sdk'
+import { sendAnalyticsEvent } from '@faststore/sdk'
 import { SearchInput as UISearchInput } from '@faststore/ui'
 import { useRouter } from 'next/router'
 import {
   forwardRef,
   lazy,
   Suspense,
-  useCallback,
   useRef,
   useState,
+  useDeferredValue,
 } from 'react'
 import type { SearchEvent } from '@faststore/sdk'
 import type {
@@ -21,6 +17,11 @@ import type {
 
 import Icon from 'src/components/ui/Icon'
 import useSearchHistory from 'src/sdk/search/useSearchHistory'
+import {
+  formatSearchPath,
+  SearchInputProvider,
+} from 'src/sdk/search/useSearchInput'
+import type { SearchInputContextValue } from 'src/sdk/search/useSearchInput'
 import useOnClickOutside from 'src/sdk/ui/useOnClickOutside'
 
 const Suggestions = lazy(() => import('src/components/search/Suggestions'))
@@ -30,30 +31,11 @@ declare type SearchInputProps = {
   buttonTestId?: string
 } & Omit<UISearchInputProps, 'onSubmit'>
 
-const useSearchHandler = (callback: (term: string) => void) => {
-  const router = useRouter()
-  const { addToSearchHistory } = useSearchHistory()
-
-  return useCallback(
-    (term: string) => {
-      const { pathname, search } = formatSearchState(
-        initSearchState({
-          term,
-          base: '/s',
-        })
-      )
-
-      sendAnalyticsEvent<SearchEvent>({
-        name: 'search',
-        params: { search_term: term },
-      })
-
-      addToSearchHistory(term)
-      callback(term)
-      router.push(`${pathname}${search}`)
-    },
-    [addToSearchHistory, callback, router]
-  )
+const sendAnalytics = async (term: string) => {
+  sendAnalyticsEvent<SearchEvent>({
+    name: 'search',
+    params: { search_term: term },
+  })
 }
 
 const SearchInput = forwardRef<SearchInputRef, SearchInputProps>(
@@ -62,12 +44,19 @@ const SearchInput = forwardRef<SearchInputRef, SearchInputProps>(
     ref
   ) {
     const [searchQuery, setSearchQuery] = useState<string>('')
+    const searchQueryDeferred = useDeferredValue(searchQuery)
     const [suggestionsOpen, setSuggestionsOpen] = useState<boolean>(false)
     const searchRef = useRef<HTMLDivElement>(null)
-    const handleSearch = useSearchHandler((term: string) => {
-      setSuggestionsOpen(false)
-      setSearchQuery(term)
-    })
+    const { addToSearchHistory } = useSearchHistory()
+    const router = useRouter()
+
+    const onSearchInputSelection: SearchInputContextValue['onSearchInputSelection'] =
+      (term, path) => {
+        addToSearchHistory({ term, path })
+        sendAnalytics(term)
+        setSuggestionsOpen(false)
+        setSearchQuery(term)
+      }
 
     useOnClickOutside(searchRef, () => setSuggestionsOpen(false))
 
@@ -77,29 +66,36 @@ const SearchInput = forwardRef<SearchInputRef, SearchInputProps>(
         data-store-search-input-wrapper
         data-store-search-input-dropdown-open={suggestionsOpen}
       >
-        <UISearchInput
-          ref={ref}
-          icon={
-            <Icon
-              name="MagnifyingGlass"
-              onClick={onSearchClick}
-              data-testid={buttonTestId}
-            />
-          }
-          placeholder="Search everything at the store"
-          onChange={(e) => setSearchQuery(e.target.value)}
-          onSubmit={handleSearch}
-          onFocus={() => setSuggestionsOpen(true)}
-          value={searchQuery}
-          {...props}
-        />
-        {suggestionsOpen && (
-          <Suspense fallback={null}>
-            <div data-store-search-input-dropdown-wrapper>
-              <Suggestions term={searchQuery} onSearch={handleSearch} />
-            </div>
-          </Suspense>
-        )}
+        <SearchInputProvider onSearchInputSelection={onSearchInputSelection}>
+          <UISearchInput
+            ref={ref}
+            icon={
+              <Icon
+                name="MagnifyingGlass"
+                onClick={onSearchClick}
+                data-testid={buttonTestId}
+              />
+            }
+            placeholder="Search everything at the store"
+            onChange={(e) => setSearchQuery(e.target.value)}
+            onSubmit={(term) => {
+              const path = formatSearchPath(term)
+
+              onSearchInputSelection(term, path)
+              router.push(path)
+            }}
+            onFocus={() => setSuggestionsOpen(true)}
+            value={searchQuery}
+            {...props}
+          />
+          {suggestionsOpen && (
+            <Suspense fallback={null}>
+              <div data-store-search-input-dropdown-wrapper>
+                <Suggestions term={searchQueryDeferred} />
+              </div>
+            </Suspense>
+          )}
+        </SearchInputProvider>
       </div>
     )
   }