+++ packages/core/src/components/search/History/SearchHistory.stories.tsx
@@ -1,3 +1,5 @@
+import { SearchInputProvider } from 'src/sdk/search/useSearchInput'
+
 import { SearchHistory } from '.'
 import type { SearchHistoryProps } from '.'
 
@@ -6,23 +8,32 @@ const meta = {
   title: 'Organisms/Search/History',
 }
 
-const Template = (props: SearchHistoryProps) => (
-  <div
-    style={{
-      maxWidth: '600px',
-      margin: '0 auto',
-      padding: '0 16px',
-      background: 'white',
-    }}
-  >
-    <SearchHistory {...props} />
-  </div>
-)
+const Template = (props: SearchHistoryProps) => {
+  return (
+    <div
+      style={{
+        maxWidth: '600px',
+        margin: '0 auto',
+        padding: '0 16px',
+        background: 'white',
+      }}
+    >
+      <SearchInputProvider>
+        <SearchHistory {...props} />
+      </SearchInputProvider>
+    </div>
+  )
+}
 
 export const Default = Template.bind({})
 
 Default.args = {
-  history: ['headphone', 'audio & video', 'mh-7000', 'jbl go'],
+  history: [
+    { term: 'headphone', path: '/' },
+    { term: 'audio & video', path: '/' },
+    { term: 'mh-7000', path: '/' },
+    { term: 'jbl go', path: '/' },
+  ],
 }
 
 Default.parameters = {