+++ packages/core/src/components/ui/Search/Suggestions/Suggestions.stories.tsx
@@ -1,5 +1,7 @@
 import { SessionProvider } from '@faststore/sdk'
 
+import { SearchInputProvider } from 'src/sdk/search/useSearchInput'
+
 import Suggestions from '.'
 import type { SuggestionsProps } from '.'
 
@@ -49,7 +51,9 @@ const Template = (props: SuggestionsProps) => (
     }}
   >
     <SessionProvider>
-      <Suggestions {...props} />
+      <SearchInputProvider>
+        <Suggestions {...props} />
+      </SearchInputProvider>
     </SessionProvider>
   </div>
 )
@@ -58,7 +62,10 @@ export const Default = Template.bind({})
 
 Default.args = {
   term: 'Ste',
-  terms: ['Steel', 'Stellar'],
+  terms: [
+    { value: 'Steel', count: 1 },
+    { value: 'Stellar', count: 2 },
+  ],
   products: [
     product({ id: '1', name: 'Handmade Steel Towels Practical' }),
     product({ id: '2', name: 'Steel Towels' }),