+++ packages/core/src/sdk/search/useSuggestions.ts
@@ -1,0 +1,48 @@
+import { useSession } from '@faststore/sdk'
+import { gql } from '@vtex/graphql-utils'
+
+import { useQuery } from 'src/sdk/graphql/useQuery'
+import type {
+  SearchSuggestionsQueryQuery as Query,
+  SearchSuggestionsQueryQueryVariables as Variables,
+} from '@generated/graphql'
+
+const MAX_SUGGESTIONS = 5
+
+const query = gql`
+  query SearchSuggestionsQuery(
+    $term: String!
+    $selectedFacets: [IStoreSelectedFacet!]
+  ) {
+    search(first: 5, term: $term, selectedFacets: $selectedFacets) {
+      suggestions {
+        terms {
+          value
+        }
+        products {
+          ...ProductSummary_product
+        }
+      }
+    }
+  }
+`
+
+function useSuggestions(term: string, limit: number = MAX_SUGGESTIONS) {
+  const { channel, locale } = useSession()
+
+  const { data, error } = useQuery<Query, Variables>(query, {
+    term,
+    selectedFacets: [
+      { key: 'channel', value: channel ?? '' },
+      { key: 'locale', value: locale },
+    ],
+  })
+
+  return {
+    terms: (data?.search.suggestions.terms ?? []).slice(0, limit),
+    products: (data?.search.suggestions.products ?? []).slice(0, limit),
+    isLoading: !error && !data,
+  }
+}
+
+export default useSuggestions